update: moved existed solutions according to the new order of problems
