--- problems/english/3.3/problem_33_c_en.md
@@ -1,33 +1,0 @@
-## [Lengths of All Words](../../../solutions/3.3/33_c.py)
-
-You will be given a string `sentence` with words separated by spaces.
-
-Write a list comprehension to generate a list of word lengths.
-
-### Note
-
-There should be nothing in the solution except the list comprehension.
-
-### Example 1
-
-__Input__
-```plaintext
-sentence = 'Мама мыла раму'
-```
-
-__Output__
-```plaintext
-[4, 4, 4]
-```
-
-### Example 2
-
-__Input__
-```plaintext
-sentence = 'Ехали медведи на велосипеде'
-```
-
-__Output__
-```plaintext
-[5, 7, 2, 10]
-```
