+++ problems/english/3.3/problem_33_f_en.md
@@ -1,57 +1,33 @@
-## [Letter Frequency Statistics](../../../solutions/3.3/33_f.py)
+## [Lengths of All Words](../../../solutions/3.3/33_f.py)
 
-You will be given a string `text`.
+You will be given a string `sentence` with words separated by spaces.
 
-Write an expression to generate a dictionary that contains information about the frequency of letters in the given string.
-
-When analyzing, ignore case, and use the lowercase letters that appear in the string as keys of the dictionary.
+Write a list comprehension to generate a list of word lengths.
 
 ### Note
 
-There should be nothing in the solution except the expression.
+There should be nothing in the solution except the list comprehension.
 
 ### Example 1
 
 __Input__
 ```plaintext
-text = 'Мама мыла раму!'
+sentence = 'Мама мыла раму'
 ```
 
 __Output__
 ```plaintext
-{'а': 4, 'л': 1, 'м': 4, 'р': 1, 'у': 1, 'ы': 1}
+[4, 4, 4]
 ```
 
 ### Example 2
 
 __Input__
 ```plaintext
-text = '''Ехали медведи
-На велосипеде.
-
-А за ними кот
-Задом наперёд.'''
+sentence = 'Ехали медведи на велосипеде'
 ```
 
 __Output__
 ```plaintext
-{   
-    'а': 6,
-    'в': 2,
-    'д': 5,
-    'е': 7,
-    'з': 2,
-    'и': 5,
-    'к': 1,
-    'л': 2,
-    'м': 3,
-    'н': 3,
-    'о': 3,
-    'п': 2,
-    'р': 1,
-    'с': 1,
-    'т': 1,
-    'х': 1,
-    'ё': 1
-}
+[5, 7, 2, 10]
 ```
