--- problems/english/3.3/problem_33_g_en.md
@@ -1,33 +1,0 @@
-## [Divisors](../../../solutions/3.3/33_g.py)
-
-You will be given a set `numbers`.
-
-Think of an expression to generate a dictionary containing information about the divisors of each of the given numbers.
-
-### Note
-
-There should be nothing in the solution except the expression.
-
-### Example 1
-
-__Input__
-```plaintext
-numbers = {1, 2, 3, 4, 5}
-```
-
-__Output__
-```plaintext
-{1: [1], 2: [1, 2], 3: [1, 3], 4: [1, 2, 4], 5: [1, 5]}
-```
-
-### Example 2
-
-__Input__
-```plaintext
-numbers = {15, 49, 36}
-```
-
-__Output__
-```plaintext
-{15: [1, 3, 5, 15], 36: [1, 2, 3, 4, 6, 9, 12, 18, 36], 49: [1, 7, 49]}
-```
