--- problems/english/3.3/problem_33_j_en.md
@@ -1,33 +1,0 @@
-## [RLE Reversed](../../../solutions/3.3/33_j.py)
-
-You will be given a list of tuples `rle` with characters and their repetition counts.
-
-Write an expression to generate the string from which this list was derived.
-
-### Note
-
-There should be nothing in the solution except the expression.
-
-### Example 1
-
-__Input__
-```plaintext
-rle = [('a', 2), ('b', 3), ('c', 1)]
-```
-
-__Output__
-```plaintext
-'aabbbc'
-```
-
-### Example 2
-
-__Input__
-```plaintext
-rle = [('1', 1), ('0', 2), ('5', 1), ('0', 2)]
-```
-
-__Output__
-```plaintext
-'100500'
-```
