+++ problems/english/3.3/problem_33_q_en.md
@@ -1,0 +1,33 @@
+## [Multiplication Table 2.0](../../../solutions/3.3/33_q.py)
+
+You will be given a single variable `n` — the required size of the multiplication table.
+
+Write a list comprehension to generate the multiplication table.
+
+### Note
+
+There should be nothing in the solution except the list comprehension.
+
+### Example 1
+
+__Input__
+```plaintext
+n = 3
+```
+
+__Output__
+```plaintext
+[[1, 2, 3], [2, 4, 6], [3, 6, 9]]
+```
+
+### Example 2
+
+__Input__
+```plaintext
+n = 4
+```
+
+__Output__
+```plaintext
+[[1, 2, 3, 4], [2, 4, 6, 8], [3, 6, 9, 12], [4, 8, 12, 16]]
+```
