--- problems/russian/3.3/problem_33_b_ru.md
@@ -1,33 +1,0 @@
-## [Таблица умножения 2.0](../../../solutions/3.3/33_b.py)
-
-Вашему решению будет предоставлена единственная переменная `n` — необходимый размер таблицы умножения.
-
-Напишите списочное выражения для генерации таблицы умножения.
-
-### Примечание
-
-В решении не должно быть ничего, кроме списочного выражения.
-
-### Пример 1
-
-__Ввод__
-```plaintext
-n = 3
-```
-
-__Вывод__
-```plaintext
-[[1, 2, 3], [2, 4, 6], [3, 6, 9]]
-```
-
-### Пример 2
-
-__Ввод__
-```plaintext
-n = 4
-```
-
-__Вывод__
-```plaintext
-[[1, 2, 3, 4], [2, 4, 6, 8], [3, 6, 9, 12], [4, 8, 12, 16]]
-```
