--- problems/russian/3.3/problem_33_c_ru.md
@@ -1,33 +1,0 @@
-## [Длины всех слов](../../../solutions/3.3/33_c.py)
-
-Вашему решению будет предоставлена строка `sentence` слов, разделённых пробелами.
-
-Напишите списочное выражения для генерации списка длин слов.
-
-### Примечание
-
-В решении не должно быть ничего, кроме списочного выражения.
-
-### Пример 1
-
-__Ввод__
-```plaintext
-sentence = 'Мама мыла раму'
-```
-
-__Вывод__
-```plaintext
-[4, 4, 4]
-```
-
-### Пример 2
-
-__Ввод__
-```plaintext
-sentence = 'Ехали медведи на велосипеде'
-```
-
-__Вывод__
-```plaintext
-[5, 7, 2, 10]
-```
