+++ problems/russian/3.3/problem_33_f_ru.md
@@ -1,10 +1,8 @@
-## [Буквенная статистика](../../../solutions/3.3/33_f.py)
+## [Длины всех слов](../../../solutions/3.3/33_f.py)
 
-Вашему решению будет предоставлена строка `text`.
+Вашему решению будет предоставлена строка `sentence` слов, разделённых пробелами.
 
-Напишите выражение для генерации словаря, который содержит информацию о частоте употребления букв в заданной строке.
-
-При анализе не учитывайте регистр, а ключами словаря сделайте использованные в строке буквы в нижнем регистре.
+Напишите списочное выражения для генерации списка длин слов.
 
 ### Примечание
 
@@ -14,44 +12,22 @@
 
 __Ввод__
 ```plaintext
-text = 'Мама мыла раму!'
+sentence = 'Мама мыла раму'
 ```
 
 __Вывод__
 ```plaintext
-{'а': 4, 'л': 1, 'м': 4, 'р': 1, 'у': 1, 'ы': 1}
+[4, 4, 4]
 ```
 
 ### Пример 2
 
 __Ввод__
 ```plaintext
-text = '''Ехали медведи
-На велосипеде.
-
-А за ними кот
-Задом наперёд.'''
+sentence = 'Ехали медведи на велосипеде'
 ```
 
 __Вывод__
 ```plaintext
-{   
-    'а': 6,
-    'в': 2,
-    'д': 5,
-    'е': 7,
-    'з': 2,
-    'и': 5,
-    'к': 1,
-    'л': 2,
-    'м': 3,
-    'н': 3,
-    'о': 3,
-    'п': 2,
-    'р': 1,
-    'с': 1,
-    'т': 1,
-    'х': 1,
-    'ё': 1
-}
+[5, 7, 2, 10]
 ```
