--- problems/russian/3.3/problem_33_g_ru.md
@@ -1,33 +1,0 @@
-## [Делители](../../../solutions/3.3/33_g.py)
-
-Вашему решению будет предоставлено множество `numbers`.
-
-Продумайте выражение для генерации словаря содержащего информацию о делителях каждого из заданных чисел.
-
-### Примечание
-
-В решении не должно быть ничего, кроме списочного выражения.
-
-### Пример 1
-
-__Ввод__
-```plaintext
-numbers = {1, 2, 3, 4, 5}
-```
-
-__Вывод__
-```plaintext
-{1: [1], 2: [1, 2], 3: [1, 3], 4: [1, 2, 4], 5: [1, 5]}
-```
-
-### Пример 2
-
-__Ввод__
-```plaintext
-numbers = {15, 49, 36}
-```
-
-__Вывод__
-```plaintext
-{15: [1, 3, 5, 15], 36: [1, 2, 3, 4, 6, 9, 12, 18, 36], 49: [1, 7, 49]}
-```
