--- problems/russian/3.3/problem_33_j_ru.md
@@ -1,33 +1,0 @@
-## [RLE наоборот](../../../solutions/3.3/33_j.py)
-
-Вашему решению будет предоставлен список кортежей `rle` с символами и количеством их повторений.
-
-Напишите выражение для генерации строки, из которой был получен данный список.
-
-### Примечание
-
-В решении не должно быть ничего, кроме списочного выражения.
-
-### Пример 1
-
-__Ввод__
-```plaintext
-rle = [('a', 2), ('b', 3), ('c', 1)]
-```
-
-__Вывод__
-```plaintext
-'aabbbc'
-```
-
-### Пример 2
-
-__Ввод__
-```plaintext
-rle = [('1', 1), ('0', 2), ('5', 1), ('0', 2)]
-```
-
-__Вывод__
-```plaintext
-'100500'
-```
