+++ problems/russian/3.3/problem_33_o_ru.md
@@ -1,0 +1,57 @@
+## [Буквенная статистика](../../../solutions/3.3/33_o.py)
+
+Вашему решению будет предоставлена строка `text`.
+
+Напишите выражение для генерации словаря, который содержит информацию о частоте употребления букв в заданной строке.
+
+При анализе не учитывайте регистр, а ключами словаря сделайте использованные в строке буквы в нижнем регистре.
+
+### Примечание
+
+В решении не должно быть ничего, кроме списочного выражения.
+
+### Пример 1
+
+__Ввод__
+```plaintext
+text = 'Мама мыла раму!'
+```
+
+__Вывод__
+```plaintext
+{'а': 4, 'л': 1, 'м': 4, 'р': 1, 'у': 1, 'ы': 1}
+```
+
+### Пример 2
+
+__Ввод__
+```plaintext
+text = '''Ехали медведи
+На велосипеде.
+
+А за ними кот
+Задом наперёд.'''
+```
+
+__Вывод__
+```plaintext
+{   
+    'а': 6,
+    'в': 2,
+    'д': 5,
+    'е': 7,
+    'з': 2,
+    'и': 5,
+    'к': 1,
+    'л': 2,
+    'м': 3,
+    'н': 3,
+    'о': 3,
+    'п': 2,
+    'р': 1,
+    'с': 1,
+    'т': 1,
+    'х': 1,
+    'ё': 1
+}
+```
