--- solutions/3.3/33_b.py
@@ -1,1 +1,0 @@
-[[i * j for j in range(1, n + 1)] for i in range(1, n + 1)]  # noqa F821
