--- solutions/3.3/33_c.py
@@ -1,1 +1,0 @@
-[len(word) for word in sentence.split()]  # noqa F821
