+++ solutions/3.3/33_f.py
@@ -1,5 +1,1 @@
-{
-    char: text.lower().count(char)  # noqa F821
-    for char in text.lower()  # noqa F821
-    if char.isalpha()
-}
+[len(word) for word in sentence.split()]  # noqa F821
