--- solutions/3.3/33_j.py
@@ -1,1 +1,0 @@
-"".join(char * repeats for char, repeats in rle)  # noqa F821
