+++ solutions/3.3/33_o.py
@@ -1,0 +1,5 @@
+{
+    char: text.lower().count(char)  # noqa F821
+    for char in text.lower()  # noqa F821
+    if char.isalpha()
+}
