+++ solutions/3.3/33_p.py
@@ -1,0 +1,1 @@
+"".join(char * repeats for char, repeats in rle)  # noqa F821
