+++ solutions/3.3/33_q.py
@@ -1,0 +1,1 @@
+[[i * j for j in range(1, n + 1)] for i in range(1, n + 1)]  # noqa F821
