+++ solutions/3.3/33_r.py
@@ -1,0 +1,4 @@
+{
+    num: [i for i in range(1, num + 1) if not num % i]
+    for num in numbers  # noqa F821
+}
